pyproject.toml
[project]
name = "HSE-FastAPI-ShortURL"
version = "0.1.1"
description = "ShortURL"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.41.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "asyncpg>=0.31.0",
    "alembic>=1.18.4",
    "redis>=7.3.0",
    "pydantic-settings>=2.13.1",
    "PyJWT>=2.12.0",
    "pwdlib[bcrypt]>=0.3.0",
    "python-multipart>=0.0.22",
    "email-validator>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
    "pytest-mock>=3.15.1",
    "coverage>=7.13.4",
    "pytest-xdist>=3.8.0",
    "locust>=2.43.3",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = [
    ".", "src"
]
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
